requirements.txt
requests
pandas
openpyxl
pycountry
gspread
oauth2client
pymongo
python-dotenv 
google-auth
tabulate